system_deps

Trait PathOrList

pub trait PathOrList<T> {
    // Required method
    fn join_paths(&self, other: impl AsRef<[PathBuf]>) -> OsString;
}

A trait that can represent both a reference to a Path like object or a list of paths. Used in Library::wrap_pkg_config and Library::from_internal_pkg_config to specify the list of pkg-config paths that should take priority.

Required Methods

fn join_paths(&self, other: impl AsRef<[PathBuf]>) -> OsString

Creates an string of paths appropiately joined for an environment variable. The paths in self will go before the paths in other.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety".